fix: set `onlyAutoAttachExplicit` to true by default
This is a new setting that configures whether we'll _only_ attach to
processes that explicitly/already have `--inspect`.

With the new magical js-debug powers, we set auto attach to attach to
every Node process. Users with auto attach turned on in a previous
release, however, suddenly were debugging a lot of things that they
didn't want to (e.g. git precommit hooks or the `code` CLI). For myself
I had gotten used to toggling auto attach when I wanted to, but this is
a muscle that existing users don't have.

In this change we set `onlyAutoAttachExplicit` to be on by default,
matching the previous auto attach behavior. Users who want the to attach
to everything can toggle it off -- or use the debug terminal.