fix: skipFiles not applying to files early on in the launch
Setting blackbox patterns is asynchronous to when the source is loaded,
so if the user asks to pause on exceptions the runtime may pause in a
place where we don't want it to. Double check at this point and manually
resume debugging for handled exceptions. This implementation seems to
 work identically to blackboxing (test cases represent this):

- ✅ An error is thrown and caught within skipFiles. Resumed here.
- ✅ An uncaught error is re/thrown within skipFiles. In both cases the
     stack is reported at the first non-skipped file is shown.
- ✅ An error is thrown from skipFiles and caught in user code. In both
     blackboxing and this version, the debugger will not pause.
- ✅ An error is thrown anywhere in user code. All good.

This old debugger had some better behavior here to work well in most
cases, but I think this implementation is more correct.

Fixes https://github.com/microsoft/vscode-js-debug/issues/644