feat: Create the VS Debug Server mode
This change is part of an effort for the VS implementation to move away from flat-session mode, and to as closely as possible follow the VS Code implementation, while sharing as much code as possible to avoid breakages due to divergent implementations.

This change moves most of the debug server creation logic out of vsCodeSessionManager and in to its own class (serverSessionManager), and adds the vsDebugServer entry-point which hosts a debug server without any VS Code API dependencies.