refactor: allow wd to drain connection to js-debug, perf
Fixes: https://github.com/microsoft/vscode-js-debug/issues/695

Previously we were closing the watchdog immediately when the target
process exited, but instead we should wait for the websocket net
data to drain and be sent back to js-debug.

This also implements some significant performance improvements for
logging. Mainly, tries to avoid async/await in hot paths and adds a
happy check path for some expensive regex compilation and console
formatting logic.