feat: reload sourcemap on new targets to deal with nodemon etc.
Life is a little difficult dealing with servers like nodemon or
`nest:dev`. These restart and can trigger things to re-compile on the
fly, which means the sourcemaps we went to the trouble of parsing
earlier might be invalid now. But we also don't want to just re-predict
on every single new target, since doing so can be expensive.

This implements a heuristic which I think should cover these cases well:

  - When a session's child exits (e.g. nodemon restarts a running
    process), then the next child that session spawns will receive a
    new instance of the breakpoint predictor.
  - The source map cache will now take into account source map mtimes
    to reload a map if it's discovered with a new mtime (by a breakpoint
    predictor--source maps loaded at runtime have no mtime). Also,
    after children are destroyed we will re-load source maps that lack
    an mtime, if we need to load them again.

Fixes #711