fix: process.stderr being incorrectly destroyed
This was intended to destroy a pipe reader, but that reader was removed
at some point and this code was left in there. Interestingly it does
not seem to cause an issue for most simple program, so I didn't catch
this, but causes 'random' failures later on.

Fixes https://github.com/microsoft/vscode/issues/106484