fix: use --inspect-brk-extension for the extension host
Fixes https://github.com/microsoft/vscode/issues/99222

node-debug2 did this, but at some point it was removed from js-debug.
While fixing this I found an interesting issue: the extension host
window was getting orphaned. I spent an hour digging around in vscode
to see what was different, before looking at it from the other
direction -- turns out that `process` was defined, but `process.pid`
wasn't, so we were returning telemetry but failing to get the
process ID of the extension host to terminate.