feat: improve handling of symbolic links
Fixes #776

Setting a breakpoint in a symlinked file will now set it in both its
realpath and linked path. However, if the breakpoint is hit it'll be in
the linked path. I would like to be able to map arbitrary paths back
into any locations in which they might be linked in the workspace.

The obvious approach is to introduce a scanning step for sources that
get all of their `realpath`'s, or builds a mapping of links (which I
could then efficiently search whenever a new script is parsed). This is
possible, but I would really rather avoid more scanning, especially for
logic that most users usually won't need. We _could_ have it behind a
flag but this introduces the discoverability issue we currently have.

So for now I have just added a warning that will be logged when a
separate realPath is detect and `--preserve-symlinks` is not present
in the `runtimeArgs`.

I also did some cleanup of dependencies instead of adding a ton of
wiring to the NodeSourcePathResolver. This seems to work but will need
to keep an eye out for any possible issues.