use ioc and a more CDP-like protocol
I adjusted some of the protocol. Previously we 'wrapped' CDP in our
custom messages. I think layering a second custom protocol on here is
not that useful and can make life harder for implementors.

Instead, the websocket exposes a literal CDP protocol, but with an
extra `JsDebug` domain that exposes the `subscribe` method. Usage
looks something like this:

```
Connected (press CTRL+C to quit)
> {"method":"Debugger.somethingInvalid", "params": {}, "id": 42}
< {"id":42,"error":{"code":-32601,"message":"'Debugger.somethingInvalid' wasn't found"}}

> {"method":"JsDebug.subscribe", "params": { "events": ["Debugger.resumed"] }, "id": 42}
< {"id":42,"result":{}}

> {"method":"Debugger.resume", "params": {}, "id": 42}
< {"method":"Debugger.resumed","params":{}}
< {"id":42,"result":{}}

> {"method":"JsDebug.subscribe", "params": { "events": ["Debugger.*"] }, "id": 42}
< {"id":42,"result":{}}

> {"method":"Debugger.resume", "params": {}, "id": 42}
< {"id":42,"result":{}}
< {"method":"Debugger.resumed","params":{}}
< {"method":"Debugger.resumed","params":{}}
< {"method":"Debugger.paused","params":{"callFrames":[...
```

To makes these calls and get response errors, I now expose the
`CDPSession` object on the `Cdp.Api`. This is pretty low level, but it
solves the problem of error handling well.

I added the "wildcard" event subscribing mentioned before. I also
removed the command from the UI since it's something that only
extensions will call. For testing I made a simple extension to run this
command. [1]

Later on (not in this PR) I want to have smarter domain enablement
handling. For example, when profiling we disable the debugger. There
should be some prioritization of enablements and disablements.

1. https://github.com/connor4312/cdp-proxy-requestor