Ensure that the debugAdapter's _onInitialize() is called
When launching a `pwa-node` process, the debugAdapter's _onInitialize()
method is never called, since the dap registers a listener for the
`'initialize'` message _after_ it has arrived. This causes the Thread's
`dap` to never be resolved, so any thread events (e.g. `'paused'`) are
never sent to the client! This in turn causes VSCode to never realize
that the process is stopped, so the debuggee just waits there,
patiently, forever.

This change stores the `Dap.InitializeParams` on the Binder and then
when the Thread is created, the params are forwarded so that the
DebugAdapter can _onInitialize() correctly and the thread-related events
can flow.