fix: urlToRegex related tests are failing
Two of the tests seemed to be incorrect, but were passing because of a previous bug, so this commit changes those two tests.
Another test was failing because the previous commit incorrectly disabled conversion of the whole path to a case insensitive regex pattern
on case insensitive platforms, so this commit restores that behavior while still correctly ignoring portions of the path that were already
converted to a regex pattern.