feat: disable pause while loading source maps if all breakpoints have been disabled
The current behavior is that the debugger is configured to pause while loading source maps if at least one breakpoint has been set during a debugging session. This has the side effect of switching to the sources tab on page reload even if all breakpoints have since been disabled. This commit removes the automatic pausing of the debugger while loading source maps if there are no longer any active breakpoints. Pausing while source maps are loading should resume if any breakpoints are reenabled or new breakpoints are added. This allows developers keep the benefit of more reliable breakpoints early in program without the drawback of extra pauses when they are not needed.