fix: debugged child processes in ext host causing teardown
Fixes #1289

We weren't properly attributing child processes in the extension host to
the extension host process that created them. This caused them all to
be treated as if they _were_ the extension host process, so closing
any one of them resulted in the debugger thinking the extension host had
gone away and clearing up all sessions.