fix: improve resolution in Webpack HMR
HMR "replaces" sources at runtime with new sources. V8 doesn't ever
unload parsed scripts, so they stick around but are unused. However, the
debugger wasn't aware of this, so if setting a breakpoint at a given
line it could end up displaying it at the wrong place if using an
outdated version of the file. (The breakpoint would still generally get
hit, since we resolve it to _all_ possible locations, but its display
is wrong).

Now, we add logic that identified webpack "hot reload" scripts. When
they come in, following webpack's logic of replacing any sourcemapped
scripts js-debug tracks with the one from the HMR bundle.