fix: dynamic debug configs sometimes having the wrong type
I added logic to rename the dynamic type to the "preferred" type (such
as from pwa-node to node), but didn't take into account that sometimes
the config providers return a different type to their main type. For
npm scripts, for example, they return a `node-terminal` type instead of
`pwa-node`. Changing the type to `node` resulted in broken configs.