feat: support nested sourcemaps
(This was entered as a bug, hence my investigation in debt week.)

A fairly common scenario is, like in the original issue, someone
building a bundled (e.g. with Webpack) application that includes
libraries whose sources also exist on disk. In this case, Webpack would
include the compiled library files, but not follow the sourcemaps those
libraries have for its made bundle sourcemap--unless the source-map-loader
is used, which often is not easy since webpack config customization is
not always exposed nicely.

This change supports these so-called 'recursive' sourcemaps:

- The breakpoint predictor will try to recursively resolve sourcemaps
- At runtime, if a loaded script has a source map that includes source
  content (the default mode of the webpack-dev-server) and that content
  has a sourceMappingURL, then we follow it as well.

This also led me to fix a bug in the breakpoint predictor, which you
can see in the code review...