chore: loosen restriction around enabling auto attach
Now that Node 12 is EOL and its usage is continuing to decrease in VS
Code, I think we can loosen the restriction to allow turning auto attach
on with spaces in the path.

Previously we would not allow turning on auto attach if there were
spaces in the path, and the Node.js version could not be determined
exactly (#708).

Fixes https://github.com/microsoft/vscode-js-debug/issues/1392