fix: improve skipfile performance
Previously, `skipFiles` applied a new regex and eventual call to
`Debugger.setBlackboxPatterns` for every source. In #1179, this resulted
in very bad performance, since a regex was sent for each file loaded in
the user's `node_modules` (several thousand). We needed to do this since
`setBlackboxPatterns` doesn't have a "negation", so we had to compute on
the debugger side which files satisfied `skipFiles` and send those to
the runtime.

This commit makes some things smarter, and some things less smart. It
restricts `skipFiles` to very simple globs, allowing only wildcards and
negations. But, in exchange, it's able to make a finite number of
regexes based solely on those patterns. In most cases, this means that
we never need to send further `setBlackboxPatterns` updates to the
runtime after loading.

For example,

```js
{ skipFiles: ['**/node_modules/**', '!**/node_modules/my-library/**'] }
```

...compiles to...

```
/^(?!.*\/node_modules\/my-library\/.*$).*\/node_modules\/.*$/
```

We do still use the old strategy if a URL isn't skipped but maps to a
path that does skip. We could eventually make this smarter if the
SourcePathResolver's `absolutePathToUrlRegex` was mixed in somehow...
but those can't be combined or negated as simply as glob patterns.

This does run the risk of breaking users who use advanced glob patterns
in their `skipFiles`, but I haven't really encountered any of those and
I believe the tradeoff is worth it.