fix: skipFiles creating catastrophic regexes
The default skipFiles for extension host debugging create a pair of
regexes, which compile down to something like:

```
/^(.+(\\/|$))*node_modules\\.asar\\/(.+(\\/|$))*$/i
```

The former part of this regex is a catastrophic lookahead, taking ~250ms
to test against the test string.

This fixes that and also adds a sanity unit test to ensure regexes run
reasonably fast.