fix: skipFiles being slow in remotes
Fixes https://github.com/microsoft/vscode/issues/170412

We were not rebasing remote paths correctly in the initial launch,
which caused the fallback logic to kick in and manually apply skipFiles
to every single script (and recompile globs), which was quite slow.

And add a test that I wrote while barking up the wrong tree.