refactor: improve and order handling in the `Binder` (#1619)
* refactor: improve and order handling in the `Binder`

Previously the `Binder` was a little messy. The binder managed the
"tree" of debug sessions, but teardown was not handled in a particular
order. So, for example, when terminating a Node.js session with a bunch
of child processes, the parent process could be terminated and
disconnected while its children were nominally still "running" if their
async shutdown logic didn't finish yet.

Now that things are nicely ordered, I can fix https://github.com/microsoft/vscode/issues/173993

* fix multi target launch