fix: workspace detection of nightly extension not working
Previously, the package.json pulled in at the bundle step was the "nightly"-treated version, but this is no longer the case. This causes nightly to not be able to automatically determine it's running in a remote workspace, since the extensionId is wrong.

Instead, define the version, package name, and publisher as compile-time constants.

Refs https://github.com/microsoft/vscode-js-debug/issues/1628#issuecomment-1492769589