chore: 'hoist' package to reduce pipeline variance
This adds a new build mode where the built assets will replace those in
the repo. This can be used in pipelines to avoid having the "workingDir"
that doesn't exist until later.

I also looked at modifying the checked-in package.json, but this was
difficult because:

(a) we modify the _name_ for nightly builds, which includes local
    development, and having this change based on the build is noisy.
(b) there's just a lot of package.json (about 13k lines) which
    uncompressed is sizable