feat: add basic network view, support experimental networking for node
This adds a basic network tree view that shows requests and responses
in the debugee. It has a default "go to" action which opens a cURL
representation of the request, and context menu actions to open the
response body either in a text editor or the hex editor. It also has
actions to copy the URL and replay the request.

![](https://memes.peet.io/img/24-08-aa5f7f35-332f-4a60-9fbb-32ac79018b60.png)

I initially was hoping to turn this on by default for Node.js since
their networking support in 22.6.0 inspired this change, but their
functionality right now is very limited (we basically get the URL and
nothing else.) Therefore I added an option to turn it on, but it's not
on by default there.

I think I might end up toggling this off for the next stable release
for now until we can polish it some more, but it's ship it in nightly
to get some feedback!