feat: improve display of HTML in the debugger
Advertises ANSI styles as proposed in https://github.com/microsoft/debug-adapter-protocol/issues/500,
though it works without them too, just without colors!

Previously the Nodes were displayed as naive objects, so we'd just
list their properties, which was quite useless when trying to get
a handle on the DOM. Now we display their children as the primary
element display, and have "Node Attributes" in a separate section.

![](https://memes.peet.io/img/24-09-9b2b35e1-3874-4e06-825a-5c84abeeb6e4.png)

Refs https://github.com/microsoft/vscode/pull/227729