fix: don't inject NODE_OPTIONS in Deno (#2080)
* fix: don't inject NODE_OPTIONS in Deno

In Deno 2, Deno has a `process` global. However when running Deno without I/O permissions (the default), accessing process.env.* will show a permission prompt. This is confusing for users, as their code did not trigger this prompt, but VS Code did.

Because this code doesn't do anything in Deno anyway (because we do not respect `NODE_OPTIONS`), there is no point in running it in Deno at all. I disable the code path in Deno through environment detection now.

* review comment